+++ tests/testthat/test-mpm_standardize.R
@@ -2,13 +2,17 @@
 test_that("mpm_standardize works correctly", {
 
   # mpm with inter-reproductive stage
-  reproStages <- apply(mat_f_inter, 2, function(x) any(x > 0))
-  matrixStages <- c('prop', 'active', 'active', 'active', 'active')
+  repro_stages <- apply(mat_f_inter, 2, function(x) any(x > 0))
+  matrix_stages <- c('prop', 'active', 'active', 'active', 'active')
 
-  x <- mpm_standardize(mat_u_inter, mat_f_inter, reproStages = reproStages,
-                       matrixStages = matrixStages)
+  x1 <- mpm_standardize(mat_u_inter, mat_f_inter, repro_stages = repro_stages,
+                       matrix_stages = matrix_stages)
   
-  expect_type(x, "list")
-  expect_length(x, 4)
-  expect_equal(ncol(x$matA), 4)
+  expect_type(x1, "list")
+  expect_length(x1, 4)
+  expect_equal(ncol(x1$matA), 4)
+  
+  x2 <- mpm_standardise(mat_u_inter, mat_f_inter, repro_stages = repro_stages,
+                        matrix_stages = matrix_stages)
+  expect_identical(x1, x2)  # check aliasing to GB spelling
 })