+++ tests/testthat/test-standard_stages.R
@@ -2,11 +2,11 @@
 test_that("standard_stages works correctly", {
   
   ## mpm without inter-reproductive stage
-  reproStages <- apply(mat_f, 2, function(x) any(x > 0))
-  matrixStages <- c('active', 'active', 'active', 'active')
-  r <- mpm_rearrange(mat_u, mat_f, mat_c, reproStages, matrixStages)
+  repro_stages <- apply(mat_f, 2, function(x) any(x > 0))
+  matrix_stages <- c('active', 'active', 'active', 'active')
+  r <- mpm_rearrange(mat_u, mat_f, mat_c, repro_stages, matrix_stages)
 
-  x <- standard_stages(r$matF, r$reproStages, r$matrixStages)
+  x <- standard_stages(r$matF, r$repro_stages, r$matrix_stages)
   expect_type(x, "list")
   expect_length(x, 4)
   
@@ -17,12 +17,12 @@ test_that("standard_stages works correctly", {
 
   
   ## mpm with inter-reproductive stage
-  reproStages <- apply(mat_f_inter, 2, function(x) any(x > 0))
-  matrixStages <- c('prop', 'active', 'active', 'active', 'active')
-  r <- mpm_rearrange(mat_u_inter, mat_f_inter, reproStages = reproStages,
-                     matrixStages = matrixStages)
+  repro_stages <- apply(mat_f_inter, 2, function(x) any(x > 0))
+  matrix_stages <- c('prop', 'active', 'active', 'active', 'active')
+  r <- mpm_rearrange(mat_u_inter, mat_f_inter, repro_stages = repro_stages,
+                     matrix_stages = matrix_stages)
   
-  x <- standard_stages(r$matF, r$reproStages, r$matrixStages)
+  x <- standard_stages(r$matF, r$repro_stages, r$matrix_stages)
   expect_type(x, "list")
   expect_length(x, 4)
   
@@ -33,12 +33,12 @@ test_that("standard_stages works correctly", {
   
   
   ## mpm with all stages reproductive
-  reproStages <- apply(mat_f_allrep, 2, function(x) any(x > 0))
-  matrixStages <- c('active', 'active')
-  r <- mpm_rearrange(mat_u_allrep, mat_f_allrep, reproStages = reproStages,
-                     matrixStages = matrixStages)
+  repro_stages <- apply(mat_f_allrep, 2, function(x) any(x > 0))
+  matrix_stages <- c('active', 'active')
+  r <- mpm_rearrange(mat_u_allrep, mat_f_allrep, repro_stages = repro_stages,
+                     matrix_stages = matrix_stages)
   
-  x <- standard_stages(r$matF, r$reproStages, r$matrixStages)
+  x <- standard_stages(r$matF, r$repro_stages, r$matrix_stages)
   expect_type(x, "list")
   expect_length(x, 4)
   expect_true(is.na(x$propStages))
@@ -53,17 +53,17 @@ test_that("standard_stages works correctly", {
 test_that("standard_stages warns and fails gracefully", {
   
   # arguments of different dimension
-  reproStages <- apply(mat_f, 2, function(x) any(x > 0))
-  matrixStages <- c('prop', 'active', 'active')
-  expect_error(standard_stages(mat_f, reproStages, matrixStages))
+  repro_stages <- apply(mat_f, 2, function(x) any(x > 0))
+  matrix_stages <- c('prop', 'active', 'active')
+  expect_error(standard_stages(mat_f, repro_stages, matrix_stages))
   
   # matF contains NA
-  reproStages <- apply(mat_f_na, 2, function(x) any(x > 0))
-  matrixStages <- c('prop', 'active', 'active', 'active')
-  expect_error(standard_stages(mat_f_na, reproStages, matrixStages))
+  repro_stages <- apply(mat_f_na, 2, function(x) any(x > 0))
+  matrix_stages <- c('prop', 'active', 'active', 'active')
+  expect_error(standard_stages(mat_f_na, repro_stages, matrix_stages))
   
   # no reproductive stages
-  reproStages <- apply(mat_f_zero, 2, function(x) any(x > 0))
-  matrixStages <- c('active', 'active', 'active', 'active')
-  expect_error(standard_stages(mat_f_zero, reproStages, matrixStages))
+  repro_stages <- apply(mat_f_zero, 2, function(x) any(x > 0))
+  matrix_stages <- c('active', 'active', 'active', 'active')
+  expect_error(standard_stages(mat_f_zero, repro_stages, matrix_stages))
 })